Separação dos models e controllers do cancelamento
