+++ app/model/ScheduleTimeSecretaryModel.php
@@ -1,5 +1,4 @@
 <?php
-
 namespace app\model;
 
 use PDO;
@@ -47,12 +46,12 @@ class ScheduleTimeSecretaryModel extends Connect
 
             try {
                 $stmt->execute();
-                return true; // Inserção bem-sucedida!
+                return true; //Inserção bem-sucedida!
             } catch (PDOException $e) {
-                return false; // Erro ao inserir dados
+                return false;//Erro ao inserir dados
             }
         } else {
-            return false; // Horário Indisponível
+            return false;//Horario Indisponivel
         }
     }
 
@@ -67,67 +66,16 @@ class ScheduleTimeSecretaryModel extends Connect
     }
 
     public function makeUnavailable($id_horario)
-    {
-        $sql = "UPDATE horario_disponivel SET disponivel = 0 WHERE id = :id_horario";
-        $stmt = $this->pdo->prepare($sql);
-        $stmt->bindParam(':id_horario', $id_horario, PDO::PARAM_INT);
+{
+    $sql = "UPDATE horario_disponivel SET disponivel = 0 WHERE id = :id_horario";
+    $stmt = $this->pdo->prepare($sql);
+    $stmt->bindParam(':id_horario', $id_horario, PDO::PARAM_INT);
 
-        try {
-            $stmt->execute();
-            return true; // Horário tornou-se indisponível com sucesso!
-        } catch (PDOException $e) {
-            return false; // Erro ao tornar horário indisponível
-        }
+    try {
+        $stmt->execute();
+        return true;//Horário tornou-se indisponível com sucesso!
+    } catch (PDOException $e) {
+        return false;//Erro ao tornar horário indisponível
     }
-
-    public function makeAvailable($id_horario)
-    {
-        $sql = "UPDATE horario_disponivel SET disponivel = 1 WHERE id = :id_horario";
-        $stmt = $this->pdo->prepare($sql);
-        $stmt->bindParam(':id_horario', $id_horario, PDO::PARAM_INT);
-
-        try {
-            $stmt->execute();
-            return true; // Horário tornou-se disponível com sucesso!
-        } catch (PDOException $e) {
-            return false; // Erro ao tornar horário disponível
-        }
-    }
-
-    public function cancelSchedule($id)
-    {
-        $this->pdo->beginTransaction();
-        
-        try {
-            // Primeiro, obter o id_horario_disponivel relacionado
-            $stmt = $this->pdo->prepare("SELECT id_horario_disponivel FROM consulta WHERE id = :id");
-            $stmt->bindValue(':id', $id, PDO::PARAM_INT);
-            $stmt->execute();
-            $id_horario = $stmt->fetchColumn();
-
-            // Excluir a consulta
-            $stmt = $this->pdo->prepare("DELETE FROM consulta WHERE id = :id");
-            $stmt->bindValue(':id', $id, PDO::PARAM_INT);
-            $stmt->execute();
-
-            // Tornar o horário disponível novamente
-            $this->makeAvailable($id_horario);
-
-            $this->pdo->commit();
-            return true;
-        } catch (PDOException $e) {
-            $this->pdo->rollBack();
-            return false;
-        }
-    }
-
-    public function getAppointments()
-    {
-        $sql = "SELECT consulta.id, horario_disponivel.data, horario_disponivel.hora, consulta.tipo_criador
-                FROM consulta
-                INNER JOIN horario_disponivel ON consulta.id_horario_disponivel = horario_disponivel.id
-                ORDER BY horario_disponivel.data, horario_disponivel.hora";
-        $stmt = $this->pdo->query($sql);
-        return $stmt->fetchAll(PDO::FETCH_ASSOC);
-    }
+}
 }
